Use cross-browser xml document constructor
Different browsers report different constructors for XML Documents.
Chrome 45.0.2454 and Firefox 40.0.0 report `XMLDocument`,
PhantomJS 1.9.8 reports `Document`. Because of this,
the existing tests pass in PhantomJS 1.9.8 but fail when run in
Chrome or Firefox.

Update the tests to use a dummy xml document's constructor
for to use in the tests that assert that responseXML is set to an xml document.